+++ mobile/android/android-components/docs/changelog.md
@@ -35,6 +35,7 @@ permalink: /changelog/
 * **browser-toolbar**
   * Add TalkBack support for page load status.
   * Added option to add "edit actions" that will show up next to the URL in edit mode.
+  * Added option to set a listener for clicks on the site security indicator (globe / lock icon).
 
 # 0.48.0